// src/enums/routes.ts
export enum apiRoutesEnum {
  AUTH_RENEW = '/auth/access-token',
  LOGIN = 'auth/login',
  EVENTS = 'events',
  CATEGORIES = 'categories',
  COMMENTS = 'comments',
  PREDICTIONS = 'predictions',
  COMPETITIONS = 'competitions',
  LEAGUES = 'leagues',
  ROUNDS = 'rounds',
  FIXTURES = 'fixtures',
  TOKENS = 'tokens',
  REPORTS = 'reports',
  ADMIN = 'admin',
  USERS = 'users',
  ANALYTICS = 'analytics',
  COINS = 'coins',
  NFTS = 'nfts',
  POSTS = 'posts',
  POST_COMMENTS = 'post-comments',
  NAVS = 'navs',
}

export enum clientRoutesEnum {
  HOME = '/',
  NEW_HOME = '/new-home',
  DETAIL_EVENT = '/detail-event/:eventId',
  HOST_PREDICTION = '/host-prediction',
  PREDICT_HISTORY = '/prediction-history',
  HOST_PREDICTION_DETAIL = '/host-prediction-detail/:eventId',
  PRO_EVENT_PREDICTION = '/pro-event-prediction',
  ADMIN_LOGIN = '/admin-login',
  NFT_COLLECTIOM = '/nft-collection',
  NFT_MARKET = '/nft-market',
  ADMIN = '/admin',
  DECENTRALIZED_POOL = '/decentralized-pool',
  INTRODUCE = '/introduce',
  AUDIT = '/audit',
  OUR_TEAM = '/our-team',
  DISCLAIMER = '/disclaimer',
  PRIVACY = '/privacy',
  SHARE_PREDICT = '/detail-event/:eventId/:type/:predictId',
  TERM = '/term',
  ROADMAP = '/roadmap',
  PARTNER = '/partner',
  HOST_INFO = '/host-info/:hostAddress',
  DETAIL_NFT = '/nft-detail/:nftId',
  CASHOUT_NFT = '/cashout-nft/:nftId',
  MY_INVESTMENT = '/my-investment',
  SELL_NFT = '/sell-nft',
  WORLD_CUP = '/world-cup',
  GROUPS = '/groups/:groupId',
  WIDGET_DETAIL = '/widget/:eventId/:themeWidget',
}